Prevent reading __proto__
Safari has a bug where calling `super.fn()` sends a `__proto__` key read
throw the proxy. Other browsers do not do this. But the implication is
that we will try to set up an event listener on `__proto__` on a
non-element which throws.

The tradeoff here is to either add a O(1) test for __proto__ on every
read, or instead change can-reflect to use try/catch around
addEventListener. I am conflicted on which is the better solution. For
now I'm going with this.